ignoring LS version and moved all file system scrips to a separate ts file, still getting duplicate on download
